feat: adds redirection for merged discussions
This adds automatic redirection using a middleware to merge-discussions so that whenever merged discussions are deleted, hitting them won't cause a 404 but a 301 redirection to the new discussion. This change introduces a new migration, model and middleware. I've also patched the merge discussion handler to be somewhat cleaner and more performant.

fixes #21
ref: blomstra/27yv3eb