Dk/redirections (#33)
* feat: adds redirection for merged discussions

This adds automatic redirection using a middleware to merge-discussions so that whenever merged discussions are deleted, hitting them won't cause a 404 but a 301 redirection to the new discussion. This change introduces a new migration, model and middleware. I've also patched the merge discussion handler to be somewhat cleaner and more performant.

fixes #21
ref: blomstra/27yv3eb

* fix class arguments error

* fixed the last bugs

* Apply fixes from StyleCI

* Double renumber to avoid integrity constraint issues, don't merge discussionTagged posts

* Apply fixes from StyleCI

Co-authored-by: StyleCI Bot <bot@styleci.io>
Co-authored-by: Ian Morland <ian@morland.me>